Add support for compile-time variable allocation
Add support for allocating variables at compile-time and binding them
at runtime.

This renames bytecode to Code and the compile function to Compile so
that both can be exported. In turn, an additional varidiac argument has
been added to Compile to pass in the names of variables to be bound at
runtime. This requires adding the same variadic argument to (*Code).Run
to accept the variables' values.

Variable binding is mostly hacked in by inserting a number of store
opcodes at the beginning of the bytecode and ensuring that the values
provided at runtime are pushed after the main value. As such, Code now
has a vars field to identify every variable it requires to run.

Rudimentary support for this has been added to the cli package.