Remove side-effects from array addition
When two arrays are added, gojq modifies the left-hand array, resulting
in any prior capture of the array to be modified as well if it happens
inside the capacity of the array (seen by taking the input array,
slicing it, and appending to it). Fix this copying the left-hand array
and appending to the copy. To avoid a copy when one doesn't need to
happen, only concatenate arrays when they both have non-zero length.

Includes a test case under cli that catches the previous behavior.