fix multiplying strings by infinite
Relying on the result of `int(math.Inf(1))` is not correct as it's not
guaranteed to give an expected result, which depends on the platform.

mips64 (softfloat): 0
ppc64le: 9223372036854775807
s390x:   9223372036854775807
armv7:   9223372036854775807
aarch64: 9223372036854775807
x86_64:  -9223372036854775808

This caused a test failure on mips64, as the result of `"string" * 0` is
"" rather than the expected null.

So we cannot rely on the result of this being negative. Instead, check
whether it's either +/- infinity, or NaN. The quickest way to do that in
go is to check `math.IsNaN(n - n)`.

Fixes #69