Fix combined --stream --slurp path slice reuse
This fixes path reuse in jsonStream that affected combined `--stream`
and `--slurp` CLI options. Previously, when both `--stream` and
`--slurp` were passed to gojq, jsonStream's iterator would reuse its
internal path slice when returning the next path in iteration. This
resulted in all paths in the slurped input having the same path. This
has workarounds, like using `tostream` instead of `--stream --slurp`,
but still ends up being a bug around path reuse.

To fix it, this adds a copyPath method to jsonStream that returns a new
slice that is a shallow copy of the path slice (which is thankfully
made of up numbers and strings, so there's no need for any work to
clone its contents).

This includes an additional test case under test.yaml to cover the
combined `--stream --slurp` options.