cli: Add Config.Run to set IO streams and args
Adds a new Config type which allows setting Stdout, Stdin, and Stderr,
and a Config.Run method that accepts the command line arguments.

The new exported APIs are,

    package cli

    type Config struct {
      Stdin  io.Reader
      Stdout io.Writer
      Stderr io.Writer
    }

    func (cfg *Config) Run(args []string) (exitCode int)

These two combined are a version of cli.Run that does not rely on
the process' global IO streams, or on os.Args.

Testing:
This updates cli_test to use Config.Run instead of setting up the cli
struct, so this exercises Config.Run for all core test cases.
This also adds tests that verify the no-op behavior of each field of
Config.

Resolves #181