fix(cli): update cli flags to match jq 1.7
the `gojq` CLI has some differences in flags compared to the current
release of `jq`.  some are clearly intentional, such as `--yaml-output`,
but others are different in name but not in behavior, which can prevent
`gojq` from being a drop-in replacement for `jq` in scripts.

this PR updates CLI flags to match `jq` 1.7.