cli: option -k to preserve the order of keys
It builds an external map from object pointer to the list of keys as seen in the
input. If an object does not change during the whole operation of gojq, cli can
find the original order of keys during encoding by looking into the map and
checking if the set of keys is the same. This works for both JSON and YAML.

Type orderedmap is used as an intermediate object created during decoding and
encoding (in case of YAML). After parsing is done, the list of keys is extracted
from orderedmap and the map is turned into a regular map[string]any. A pointer
to that map is added to a global table map[uintptr][]string which stores the
order of keys. When encoding, this data is used to put keys in the right order.