stop numeric normalization for concurrent execution
Numeric normalization was originally introduced to allow `int64` in
input values (see #36). However, this prevents concurrent execution on
the same data, and this has been a serious issue for package users.
However, this normalization was also needed to normalize `json.Number`
from `encoding/json` and `uint64` from `go-yaml`, because only `int`,
`float64`, and `*big.Int` are allowed during the execution step.

This commit removes numeric normalization and allows `json.Number`
during execution. So, you can now run multiple queries on the same data
concurrently. This fixes #212, fixes #236, closes #237, closes #274.
Also, it preserves precision of floating-point numbers, it fixes #216.
Note that `int64` and `uint64` are no longer allowed in input values.

The `go-yaml` package was unmaintained for a long time, and this was a
blocker for removing numeric normalization. However, the package was
forked to a new repository and active maintenance resumed recently. I
then forked the repository again, and fixed to meet gojq's requirements.
I changed it to emit `json.Number` for integers, and to decode map into
`map[string]any` (not `map[any]any`). As a result, the precision of
large numbers in YAML are now preserved, the precision of floating-point
numbers in map keys are preserved, and performance of parsing large YAML
contents is improved.